backend/requirements.txt
# Workspace version: 1.42.78
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ipykernel==6.29.5
fastapi==0.111.0
databutton==0.38.34
openai
beautifulsoup4
requests
python-jose[cryptography]
groq